support .jpeg now in clean, made subprocess call work, ready to try full script

+++ uncertain/imageCleaningAndGoogleSearching/clean.py
@@ -11,7 +11,7 @@ for f in anyFiles:
     relPathAndName = join(dir, f)
 
     # delete non images (and duplicates, which have a .1 or .n at end, where n is an integer)
-    if fExt != ".png" and fExt != ".jpg" and fExt != ".JPEG":
+    if fExt != ".png" and fExt != ".jpg" and fExt != ".JPEG" and fExt != ".jpeg":
         os.remove(relPathAndName)
     # otherwise, reshape them
     else:
